Bug 40933 emailin editor debug
git-svn-id: svn://svn.office.custis.ru/3rdparty/bugzilla.org/trunk@318 6955db30-a419-402b-8a0d-67ecbb4d7f56

+++ email_in.pl
@@ -185,7 +185,7 @@ sub post_bug {
     $cgi->param(-name => 'inbound_email', -value => 1);
 
     my $bug_id = do 'post_bug.cgi';
-    debug_print($@) if $@;
+    die $@ . "\n\nIncoming mail format for entering bugs:\n\@field = value\n\@field = value\n...\n\nBug text\n" if $@;
     if ($fields{attachments} && @{$fields{attachments}})
     {
         $cgi->delete(keys %fields);
